// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.16.1.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as schema from "../schema.js";
import type * as vans from "../vans.js";
import type * as jobs from "../jobs.js";
import type * as technicians from "../technicians.js";
import type * as clients from "../clients.js";
import type * as inventory from "../inventory.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  schema: typeof schema;
  vans: typeof vans;
  jobs: typeof jobs;
  technicians: typeof technicians;
  clients: typeof clients;
  inventory: typeof inventory;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;